refactor: ユーザー情報表示部分のファイル分割
