+++ src/components/index.ts
@@ -20,3 +20,4 @@ export { default as DisplayProducts } from "./common_part/DisplayProducts";
 export { default as Like } from "./common_part/Like";
 export { default as DisplayTagList } from "./Home/DisplayTagList";
 export { default as postImage } from "./common_part/postImage";
+export { default as DisplayUserInfo } from "./user/DisplayUserInfo";
